pyproject.toml
[project]
name = "e-commerce-customer-analytics-mlops-pipeline"
version = "0.1.0"
description = ""
authors = [
    {name = "mert0afacan11@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pyspark (>=4.0.0,<5.0.0)",
    "scikit-learn (==1.7.0)",
    "pandas (>=2.3.1,<3.0.0)",
    "numpy (>=2.3.2,<3.0.0)",
    "xgboost (>=3.0.3,<4.0.0)",
    "lifetimes (>=0.11.3,<0.12.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "mlflow (==3.2.0)",
    "dagshub (>=0.6.2,<0.7.0)",
    "hydra-core (>=1.3.2,<2.0.0)",
    "hydra-colorlog (>=1.2.0,<2.0.0)",
    "apache-airflow (>=3.0.4,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
